Change bonjourr collection structure
