+++ services/src/backgrounds/pixabay/images/search.ts
@@ -1,3 +1,5 @@
+import { pixabayImageToGeneric } from '../convert.ts'
+
 import type { Image, Pixabay, PixabayImage } from '../../../../types/backgrounds.ts'
 import type { Env } from '../../../index.ts'
 
@@ -12,16 +14,7 @@ export async function pixabayImagesSearch(url: URL, env: Env, headers: Headers):
 	const json = await resp.json<Pixabay>()
 
 	const arr = json.hits as PixabayImage[]
-	const result: Image[] = arr.map((item) => ({
-		format: 'image',
-		urls: {
-			full: item.largeImageURL,
-			medium: item.webformatURL,
-			small: item.previewURL,
-		},
-		page: item.pageURL,
-		username: item.user,
-	}))
+	const result: Image[] = arr.map((item) => pixabayImageToGeneric(item))
 
 	return new Response(JSON.stringify({ 'pixabay-images-search': result }), {
 		headers,
