+++ services/src/backgrounds/unsplash/images/user.ts
@@ -1,4 +1,5 @@
-import { fetchUnsplash, toBonjourrImages } from '../shared.ts'
+import { unsplashToGeneric } from '../convert.ts'
+import { fetchUnsplash } from '../shared.ts'
 
 export async function unsplashImagesSearch(url: URL, headers: Headers): Promise<Response> {
 	const orientation = url.searchParams.get('orientation') ?? 'landscape'
@@ -7,7 +8,7 @@ export async function unsplashImagesSearch(url: URL, headers: Headers): Promise<
 	const h = url.searchParams.get('h') ?? '1080'
 
 	const images = await fetchUnsplash(`?query=${query}&orientation=${orientation}&content_filter=high&count=20`)
-	const result = toBonjourrImages(images, w, h)
+	const result = images.map((image) => (unsplashToGeneric(image, w, h)))
 
 	return new Response(JSON.stringify({ 'unsplash-images-search': result }), {
 		headers,
@@ -21,7 +22,7 @@ export async function unsplashImagesCollections(url: URL, headers: Headers): Pro
 	const h = url.searchParams.get('h') ?? '1080'
 
 	const images = await fetchUnsplash(`?collections=${query}&orientation=${orientation}&count=20`)
-	const result = toBonjourrImages(images, w, h)
+	const result = images.map((image) => (unsplashToGeneric(image, w, h)))
 
 	return new Response(JSON.stringify({ 'unsplash-images-collections': result }), {
 		headers,
